+++ eslint.config.js
@@ -220,6 +220,18 @@ export default defineConfig([
     },
   },
   {
+    name: "alloy/node-globals",
+    files: [
+      "packages/node/src/**/*.{cjs,js}",
+      "packages/node/test/**/*.{cjs,js}",
+    ],
+    languageOptions: {
+      globals: {
+        ...globals.node,
+      },
+    },
+  },
+  {
     name: "alloy/core-src",
     files: ["packages/*/src/**/*.{cjs,js}"],
     rules: {
@@ -448,6 +460,12 @@ export default defineConfig([
     },
     rules: {
       "no-console": "off",
+      "import/extensions": [
+        "error",
+        {
+          js: "always",
+        },
+      ],
     },
   },
   {